Support of Faults according to SOAP 1.2 specification
Adding a way to the server to know which SOAP Fault format it should use.
Proposing a setfault function that matches faults 1.2 attributes

KDSoapFaultException detects the version according to the tagname and
make the proper deserialization.
Subcodes hierarchy handled in a helper function and the class provides
conversion from faultCode enum (and vice versa).